BLACKBOX_CODE/PRECOND_BLACKBOX/requirements.txt
atari-py==0.2.6
cloudpickle==1.3.0
cycler==0.10.0
future==0.18.2
gym==0.17.1
joblib==0.15.1
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.3
Pillow==7.1.2
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
torch==1.5.0
torchvision==0.6.0
